+++ fatiando/potential/_neprism.py
@@ -10,9 +10,7 @@ import numpy
 from numpy import sqrt, log, arctan2
 import numexpr
 
-SI2EOTVOS = 1000000000.0
-SI2MGAL = 100000.0
-G = 0.00000000006673
+from fatiando.constants import SI2EOTVOS, SI2MGAL, G
 
 
 def gz(xp, yp, zp, prisms, dens=None):